Extract apigateway to separate file
