+++ components/console-backend-service/internal/domain/apigateway.resolvers.go
@@ -1,0 +1,35 @@
+package domain
+
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+
+import (
+	"context"
+
+	"github.com/kyma-incubator/api-gateway/api/v1alpha1"
+	"github.com/kyma-project/kyma/components/console-backend-service/internal/gqlschema"
+)
+
+func (r *mutationResolver) CreateAPIRule(ctx context.Context, name string, namespace string, params v1alpha1.APIRuleSpec) (*v1alpha1.APIRule, error) {
+	return r.ag.CreateAPIRule(ctx, name, namespace, params)
+}
+
+func (r *mutationResolver) UpdateAPIRule(ctx context.Context, name string, namespace string, params v1alpha1.APIRuleSpec) (*v1alpha1.APIRule, error) {
+	return r.ag.UpdateAPIRule(ctx, name, namespace, params)
+}
+
+func (r *mutationResolver) DeleteAPIRule(ctx context.Context, name string, namespace string) (*v1alpha1.APIRule, error) {
+	return r.ag.DeleteAPIRule(ctx, name, namespace)
+}
+
+func (r *queryResolver) APIRules(ctx context.Context, namespace string, serviceName *string, hostname *string) ([]*v1alpha1.APIRule, error) {
+	return r.ag.APIRulesQuery(ctx, namespace, serviceName, hostname)
+}
+
+func (r *queryResolver) APIRule(ctx context.Context, name string, namespace string) (*v1alpha1.APIRule, error) {
+	return r.ag.APIRuleQuery(ctx, name, namespace)
+}
+
+func (r *subscriptionResolver) APIRuleEvent(ctx context.Context, namespace string, serviceName *string) (<-chan *gqlschema.APIRuleEvent, error) {
+	return r.ag.APIRuleEventSubscription(ctx, namespace, serviceName)
+}
